session: split `TestInitializeSQLFile` into four ones to speed up (#46968)
close pingcap/tidb#46832
