+++ session/test/BUILD.bazel
@@ -8,7 +8,7 @@ go_test(
         "session_test.go",
     ],
     flaky = True,
-    shard_count = 29,
+    shard_count = 33,
     deps = [
         "//config",
         "//domain",
